PanelSettings.settingsTree performance improvements (#275)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
Performance improvements mainly in layouts with 3D and Plot panels.
Consequently, the looping issue when changing the Plot Planning
parameters (ie label input) no longer occurs.


**Description**
Use the React Dev Tools highlight feature to observe components'
rendering in real time.
It's really helpful to identify each component that has been
over-rendered.

- `PanelAPI`: Performance improvements in the `useConfigById` function
using useMemo to memorize hard processing.
- `PanelSettings`: Memoized settingsTrees building to decrease
processing in each cycle.
- `PanelSettings`: Some types were moved to the proper file.
- `SettingsTreeEditor.FieldEditor`: Types were moved to the proper file.
- `SettingsTreeEditor.NodeActionsMenu`: Handle button actions (click and
close) using useCallback hook.
- `SettingsTreeEditor.NodeActionsMenu`: Style was moved to the proper
file (NodeActionsMenu.style.ts)
- `SettingsTreeEditor.NodeEditor`: Using useMemo to improve performance.
- `SettingsTreeEditor`:  Constants were moved to the proper file.
- `SettingsTreeEditor.index`: The node rendering was improved using
useMemo.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [ ] This change is covered by unit tests
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated

---------
