+++ packages/suite-base/src/components/PanelSettings/index.tsx
@@ -14,7 +14,6 @@ import { SettingsTree } from "@lichtblick/suite";
 import { AppSetting } from "@lichtblick/suite-base/AppSetting";
 import { useConfigById } from "@lichtblick/suite-base/PanelAPI";
 import { useMessagePipelineGetter } from "@lichtblick/suite-base/components/MessagePipeline";
-import { getTopicToSchemaNameMap } from "@lichtblick/suite-base/components/MessagePipeline/selectors";
 import { ActionMenu } from "@lichtblick/suite-base/components/PanelSettings/ActionMenu";
 import { EmptyWrapper } from "@lichtblick/suite-base/components/PanelSettings/EmptyWrapper";
 import { buildSettingsTree } from "@lichtblick/suite-base/components/PanelSettings/settingsTree";
@@ -134,17 +133,32 @@ export default function PanelSettings({
 
   const [config, , extensionSettings] = useConfigById(selectedPanelId);
   const messagePipelineState = useMessagePipelineGetter();
-  const topicToSchemaNameMap = getTopicToSchemaNameMap(messagePipelineState());
 
-  const settingsTree = usePanelStateStore((state) =>
-    buildSettingsTree({
+  const storedSettingsTrees = usePanelStateStore(({ settingsTrees }) => settingsTrees);
+  const settingsTree = useMemo(
+    () =>
+      buildSettingsTree({
+        config,
+        extensionSettings,
+        messagePipelineState,
+        panelType,
+        selectedPanelId,
+        settingsTrees: storedSettingsTrees,
+      }),
+    [
       config,
       extensionSettings,
+      messagePipelineState,
       panelType,
       selectedPanelId,
-      state,
-      topicToSchemaNameMap,
-    }),
+      /**
+       * The core issue is that settingsTrees object in the PanelStateStore is being
+       * mutated on each render, leading to unnecessary calls to buildSettingsTree
+       * To address this, we need to ensure that settingsTrees remains
+       * referentially stable unless its actual content changes.
+       */
+      storedSettingsTrees,
+    ],
   );
 
   const resetToDefaults = useCallback(() => {
