+++ packages/suite-base/src/components/PanelSettings/types.ts
@@ -3,6 +3,18 @@
 
 import { SvgIconProps } from "@mui/material";
 
+import { PanelSettings } from "@lichtblick/suite";
+import { MessagePipelineContext } from "@lichtblick/suite-base/components/MessagePipeline";
+import { PanelStateStore } from "@lichtblick/suite-base/context/PanelStateContext";
+
+export type BuildSettingsTreeProps = {
+  config: Record<string, unknown> | undefined;
+  extensionSettings: Record<string, Record<string, PanelSettings<unknown>>>;
+  messagePipelineState: () => MessagePipelineContext;
+  panelType: string | undefined;
+  selectedPanelId: string | undefined;
+} & Pick<PanelStateStore, "settingsTrees">;
+
 export type ActionMenuProps = {
   allowShare: boolean;
   onReset: () => void;
