+++ packages/suite-base/src/components/SettingsTreeEditor/FieldEditor.tsx
@@ -19,11 +19,11 @@ import {
   Tooltip,
   Typography,
 } from "@mui/material";
-import { makeStyles } from "tss-react/mui";
 import { v4 as uuid } from "uuid";
 
 import { Immutable, SettingsTreeAction, SettingsTreeField } from "@lichtblick/suite";
 import MessagePathInput from "@lichtblick/suite-base/components/MessagePathSyntax/MessagePathInput";
+import { useStyles } from "@lichtblick/suite-base/components/SettingsTreeEditor/FieldEditor.style";
 import Stack from "@lichtblick/suite-base/components/Stack";
 import { useAppContext } from "@lichtblick/suite-base/context/AppContext";
 
@@ -33,77 +33,6 @@ import { ColorGradientInput, ColorPickerInput, NumberInput, Vec2Input, Vec3Input
 const UNDEFINED_SENTINEL_VALUE = uuid();
 /** Used to avoid MUI errors when an invalid option is selected */
 const INVALID_SENTINEL_VALUE = uuid();
-
-const useStyles = makeStyles<void, "error">()((theme, _params, classes) => ({
-  autocomplete: {
-    ".MuiInputBase-root.MuiInputBase-sizeSmall": {
-      paddingInline: 0,
-      paddingBlock: theme.spacing(0.3125),
-    },
-  },
-  clearIndicator: {
-    marginRight: theme.spacing(-0.25),
-    opacity: theme.palette.action.disabledOpacity,
-
-    ":hover": {
-      background: "transparent",
-      opacity: 1,
-    },
-  },
-  error: {},
-  fieldLabel: {
-    color: theme.palette.text.secondary,
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-    whiteSpace: "nowrap",
-  },
-  fieldWrapper: {
-    minWidth: theme.spacing(14),
-    marginRight: theme.spacing(0.5),
-    [`&.${classes.error} .MuiInputBase-root, .MuiInputBase-root.${classes.error}`]: {
-      outline: `1px ${theme.palette.error.main} solid`,
-      outlineOffset: -1,
-    },
-  },
-  multiLabelWrapper: {
-    display: "grid",
-    gridTemplateColumns: "1fr auto",
-    columnGap: theme.spacing(0.5),
-    height: "100%",
-    width: "100%",
-    alignItems: "center",
-    textAlign: "end",
-  },
-  styledToggleButtonGroup: {
-    backgroundColor: theme.palette.action.hover,
-    gap: theme.spacing(0.25),
-    overflowX: "auto",
-
-    "& .MuiToggleButtonGroup-grouped": {
-      margin: theme.spacing(0.55),
-      borderRadius: theme.shape.borderRadius,
-      paddingTop: 0,
-      paddingBottom: 0,
-      borderColor: "transparent !important",
-      lineHeight: 1.75,
-
-      "&.Mui-selected": {
-        background: theme.palette.background.paper,
-        borderColor: "transparent",
-
-        "&:hover": {
-          borderColor: theme.palette.action.active,
-        },
-      },
-      "&:not(:first-of-type)": {
-        borderRadius: theme.shape.borderRadius,
-      },
-      "&:first-of-type": {
-        borderRadius: theme.shape.borderRadius,
-      },
-    },
-  },
-}));
 
 function FieldInput({
   actionHandler,
