+++ packages/suite-base/src/components/SettingsTreeEditor/constants.ts
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+export const NODE_HEADER_MIN_HEIGHT = 35;
